# openapi/workflow-tracker-v1.yaml
openapi: 3.0.3
info:
  title: Workflow Tracker API
  version: 1.0.0
  description: Tracks asynchronous workflow events and downstream acknowledgements.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/workflows:
    post:
      summary: Create a workflow event
      operationId: createWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '201':
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/workflows/pending:
    get:
      summary: List pending workflows
      operationId: listPendingWorkflows
      responses:
        '200':
          description: Pending workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkflowResponse'
  /api/v1/workflows/{workflowId}:
    get:
      summary: Get workflow state
      operationId: getWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: Workflow state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/workflows/{workflowId}/acknowledge:
    post:
      summary: Acknowledge a workflow from a downstream service
      operationId: acknowledgeWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcknowledgeRequest'
      responses:
        '200':
          description: Acknowledgement recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/workflows/{workflowId}/failed:
    post:
      summary: Mark an overdue workflow as failed
      operationId: failWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FailRequest'
      responses:
        '200':
          description: Workflow failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  parameters:
    WorkflowId:
      name: workflowId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    CreateWorkflowRequest:
      type: object
      required: [eventId, targetServices]
      properties:
        eventId:
          type: string
          maxLength: 100
        payload:
          type: object
        targetServices:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
    AcknowledgeRequest:
      type: object
      required: [serviceName]
      properties:
        serviceName:
          type: string
    FailRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 500
    WorkflowResponse:
      type: object
      properties:
        workflowId:
          type: string
          format: uuid
        eventId:
          type: string
        payload:
          type: object
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED]
        failureReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        ackDeadline:
          type: string
          format: date-time
        acknowledgedServices:
          type: array
          items:
            type: string
        pendingServices:
          type: array
          items:
            type: string
    Problem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Workflow not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Invalid workflow state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'